+++ flowman-plugins/swagger/src/test/scala/com/dimajix/flowman/spec/schema/SwaggerSchemaTest.scala
@@ -22,8 +22,10 @@ import org.scalatest.matchers.should.Matchers
 import com.dimajix.flowman.execution.Session
 import com.dimajix.flowman.spec.ObjectMapper
 import com.dimajix.flowman.types.ArrayType
+import com.dimajix.flowman.types.BooleanType
 import com.dimajix.flowman.types.DecimalType
 import com.dimajix.flowman.types.DoubleType
+import com.dimajix.flowman.types.Field
 import com.dimajix.flowman.types.FloatType
 import com.dimajix.flowman.types.IntegerType
 import com.dimajix.flowman.types.LongType
@@ -92,6 +94,90 @@ class SwaggerSchemaTest extends AnyFlatSpec with Matchers  {
         fields(2).name should be ("id")
         fields(2).description should be (Some("The Pets ID"))
         fields(2).ftype should be (LongType)
+
+        session.shutdown()
+    }
+
+    it should "support additionalProperties: false" in {
+        val spec =
+            """
+              |kind: swagger
+              |entity: Pet
+              |spec: |
+              |    swagger: "2.0"
+              |    info:
+              |      version: 1.0.0
+              |      title: Swagger Petstore
+              |      description: A sample API that uses a petstore as an example to demonstrate features in the swagger-2.0 specification
+              |      termsOfService: http://swagger.io/terms/
+              |      contact:
+              |        name: Swagger API Team
+              |        email: apiteam@swagger.io
+              |        url: http://swagger.io
+              |      license:
+              |        name: Apache 2.0
+              |    url: https://www.apache.org/licenses/LICENSE-2.0.html
+              |    definitions:
+              |      Pet:
+              |        required:
+              |          - name
+              |          - id
+              |          - transactionData
+              |        properties:
+              |          name:
+              |            type: string
+              |            description: The Pets name
+              |          tag:
+              |            type: string
+              |          id:
+              |            type: integer
+              |            format: int64
+              |            description: The Pets ID
+              |          transactionData:
+              |            description: 'transaction type meta-information'
+              |            type: object
+              |            properties:
+              |              flight:
+              |                description: 'transaction type flight, if transactionType is set to Flight this might contain data'
+              |                additionalProperties: false
+              |                properties:
+              |                  additionalDataOnlyIndicator:
+              |                    description: 'Indicates things'
+              |                    type: boolean
+              |                  uid:
+              |                    type: string
+              |                    description: unique ID (UUID)
+              |                    format: uuid
+              |                    example: 105a76d8-db49-4144-ace7-e683e8f4ba46
+              |""".stripMargin
+
+        val session = Session.builder().disableSpark().build()
+        val schemaSpec = ObjectMapper.parse[SchemaSpec](spec)
+        schemaSpec shouldBe an[SwaggerSchemaSpec]
+
+        val result = schemaSpec.instantiate(session.context)
+        result shouldBe an[SwaggerSchema]
+        result.description should be(Some("A sample API that uses a petstore as an example to demonstrate features in the swagger-2.0 specification"))
+
+        result.fields should be (Seq(
+            Field("name", StringType, nullable = false, description = Some("The Pets name")),
+            Field("tag", StringType, nullable = true, description = None),
+            Field("id", LongType, nullable = false, description = Some("The Pets ID"), format=Some("int64")),
+            Field("transactionData", StructType(
+                Seq(
+                    Field("flight", StructType(
+                        Seq(
+                            Field("additionalDataOnlyIndicator", BooleanType, nullable = true, description = Some("Indicates things")),
+                            Field("uid", StringType, nullable = true, format=Some("uuid"), description = Some("unique ID (UUID)"))
+                        )),
+                        nullable = true,
+                        description = Some("transaction type flight, if transactionType is set to Flight this might contain data")
+                    )
+                )),
+                nullable = false,
+                description = Some("transaction type meta-information")
+            )
+        ))
 
         session.shutdown()
     }